fix : register 로직 수정
